+++ springboot-master/springboot-master/src/main/java/at/ac/fhsalzburg/swd/spring/repository/AudioRepository.java
@@ -8,7 +8,7 @@ import at.ac.fhsalzburg.swd.spring.model.Audio;
 import at.ac.fhsalzburg.swd.spring.model.Media;
 
 @Repository
-public interface AudioRepository extends CrudRepository<Audio, Integer> {
+public interface AudioRepository extends CrudRepository<Audio, Long> {
 
 	@Transactional(timeout = 10)
 	Media findById(int id);
