+++ springboot-master/springboot-master/src/main/java/at/ac/fhsalzburg/swd/spring/repository/EditionRepository.java
@@ -10,7 +10,7 @@ import at.ac.fhsalzburg.swd.spring.model.Edition;
 import at.ac.fhsalzburg.swd.spring.model.Media;
 
 @Repository
-public interface EditionRepository extends CrudRepository<Edition, Integer>{
+public interface EditionRepository extends CrudRepository<Edition, Long>{
 
 	@Transactional(timeout = 10)
     Edition findById(int id);
