+++ springboot-master/springboot-master/src/main/java/at/ac/fhsalzburg/swd/spring/repository/MediaTransactionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MediaTransactionRepository extends CrudRepository<MediaTransaction, Integer> {
+public interface MediaTransactionRepository extends CrudRepository<MediaTransaction, Long> {
 
 	//TODO: @Transactional
 	//TODO: @Query
