+++ springboot-master/springboot-master/src/main/java/at/ac/fhsalzburg/swd/spring/repository/MediaTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import at.ac.fhsalzburg.swd.spring.model.MediaType;
 
 @Repository
-public interface MediaTypeRepository extends CrudRepository<MediaType, Integer> {
+public interface MediaTypeRepository extends CrudRepository<MediaType, Long> {
 
     @Transactional(timeout = 10)
     MediaType findById(int id);
