+++ springboot-master/springboot-master/src/main/java/at/ac/fhsalzburg/swd/spring/repository/ShelfRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import at.ac.fhsalzburg.swd.spring.model.Shelf;
 
 @Repository
-public interface ShelfRepository extends CrudRepository<Shelf, Integer> {
+public interface ShelfRepository extends CrudRepository<Shelf, Long> {
 
 	@Transactional(timeout = 10)
 	Shelf findById(int id);
